data/azure-static-web-apps-with-api/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.24.0",
]
